Added Exception Handling for Get Score and Delete Score API

+++ src/main/java/com/oyo/topscore/service/impl/TopScoreServiceImpl.java
@@ -52,7 +52,9 @@ public class TopScoreServiceImpl implements TopScoreService {
      * @return {@inheritDoc}
      */
     public String deleteScore(Long id) {
-        playerDetailRepository.delete(playerDetailRepository.getOne(id));
+        if (playerDetailRepository.findById(id).get() !=null) {
+            playerDetailRepository.delete(playerDetailRepository.getOne(id));
+        }
         return "Record with id: "+id+" is deleted";
     }
